Moved each loader to its own file
- EXPERIMENTAL
