+++ ragondin/components/__init__.py
@@ -1,6 +1,6 @@
 from config import load_config
-from .indexer import AudioTranscriber, Indexer
+from .indexer import Indexer
 from .pipeline import RagPipeline
-from .indexer import AudioTranscriber, Indexer
+from .indexer import Indexer
 
-__all__ = [load_config, RagPipeline, Indexer, AudioTranscriber]
+__all__ = [load_config, RagPipeline, Indexer]
