+++ ragondin/components/indexer/loaders/AudioTranscriber.py
@@ -1,0 +1,10 @@
+from components.utils import SingletonMeta
+import whisperx
+
+class AudioTranscriber(metaclass=SingletonMeta):
+    def __init__(self, device='cpu', compute_type='float32', model_name='large-v2', language='fr'):
+        self.model = whisperx.load_model(
+            model_name, 
+            device=device, language=language, 
+            compute_type=compute_type
+        )
